address review comments (part-2)
