+++ pkg/plugin/plugin.go
@@ -21,68 +21,37 @@ package plugin
 
 import (
 	"crypto/rand"
-	"encoding/json"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"net"
-	"os"
 	"runtime"
 	"sort"
-	"strings"
 	"time"
 
 	"github.com/containernetworking/cni/pkg/skel"
-	"github.com/containernetworking/cni/pkg/types"
+	cnitypes "github.com/containernetworking/cni/pkg/types"
 	"github.com/containernetworking/cni/pkg/types/current"
 	"github.com/containernetworking/plugins/pkg/ip"
 	"github.com/containernetworking/plugins/pkg/ipam"
 	"github.com/containernetworking/plugins/pkg/ns"
-	"github.com/imdario/mergo"
 	"github.com/j-keck/arping"
 	"github.com/vishvananda/netlink"
 
+	"github.com/k8snetworkplumbingwg/ovs-cni/pkg/config"
 	"github.com/k8snetworkplumbingwg/ovs-cni/pkg/ovsdb"
 	"github.com/k8snetworkplumbingwg/ovs-cni/pkg/sriov"
+	"github.com/k8snetworkplumbingwg/ovs-cni/pkg/types"
 	"github.com/k8snetworkplumbingwg/ovs-cni/pkg/utils"
 )
 
-const (
-	macSetupRetries        = 2
-	linkstateCheckRetries  = 5
-	linkStateCheckInterval = 600 // in milliseconds
-)
-
-type netConf struct {
-	types.NetConf
-	BrName                 string   `json:"bridge,omitempty"`
-	VlanTag                *uint    `json:"vlan"`
-	MTU                    int      `json:"mtu"`
-	Trunk                  []*trunk `json:"trunk,omitempty"`
-	DeviceID               string   `json:"deviceID"` // PCI address of a VF in valid sysfs format
-	ConfigurationPath      string   `json:"configuration_path"`
-	SocketFile             string   `json:"socket_file"`
-	LinkStateCheckRetries  int      `json:"link_state_check_retries"`
-	LinkStateCheckInterval int      `json:"link_state_check_interval"`
-}
-
-type trunk struct {
-	MinID *uint `json:"minID,omitempty"`
-	MaxID *uint `json:"maxID,omitempty"`
-	ID    *uint `json:"id,omitempty"`
-}
+const macSetupRetries = 2
 
 // EnvArgs args containing common, desired mac and ovs port name
 type EnvArgs struct {
-	types.CommonArgs
-	MAC     types.UnmarshallableString `json:"mac,omitempty"`
-	OvnPort types.UnmarshallableString `json:"ovnPort,omitempty"`
-}
-
-type cachedNetConf struct {
-	Netconf    *netConf
-	OrigIfName string
+	cnitypes.CommonArgs
+	MAC     cnitypes.UnmarshallableString `json:"mac,omitempty"`
+	OvnPort cnitypes.UnmarshallableString `json:"ovnPort,omitempty"`
 }
 
 func init() {
@@ -100,7 +69,7 @@ func logCall(command string, args *skel.CmdArgs) {
 func getEnvArgs(envArgsString string) (*EnvArgs, error) {
 	if envArgsString != "" {
 		e := EnvArgs{}
-		err := types.LoadArgs(envArgsString, &e)
+		err := cnitypes.LoadArgs(envArgsString, &e)
 		if err != nil {
 			return nil, err
 		}
@@ -116,93 +85,6 @@ func getHardwareAddr(ifName string) string {
 	}
 	return ifLink.Attrs().HardwareAddr.String()
 
-}
-
-func loadAllNetConf(data []byte) (*netConf, error) {
-	netconf, err := loadNetConf(data)
-	if err != nil {
-		return nil, err
-	}
-	flatNetConf, err := loadFlatNetConf(netconf.ConfigurationPath)
-	if err != nil {
-		return nil, err
-	}
-	netconf, err = mergeConf(netconf, flatNetConf)
-	if err != nil {
-		return nil, err
-	}
-	if netconf.LinkStateCheckRetries == 0 {
-		netconf.LinkStateCheckRetries = linkstateCheckRetries
-	}
-
-	if netconf.LinkStateCheckInterval == 0 {
-		netconf.LinkStateCheckInterval = linkStateCheckInterval
-	}
-	return netconf, nil
-}
-
-func loadNetConf(bytes []byte) (*netConf, error) {
-	netconf := &netConf{}
-	if err := json.Unmarshal(bytes, netconf); err != nil {
-		return nil, fmt.Errorf("failed to load netconf: %v", err)
-	}
-
-	return netconf, nil
-}
-
-func loadFlatNetConf(configPath string) (*netConf, error) {
-	confFiles := getOvsConfFiles()
-	if configPath != "" {
-		confFiles = append([]string{configPath}, confFiles...)
-	}
-
-	// loop through the path and parse the JSON config
-	flatNetConf := &netConf{}
-	for _, confFile := range confFiles {
-		confExists, err := pathExists(confFile)
-		if err != nil {
-			return nil, fmt.Errorf("error checking ovs config file: error: %v", err)
-		}
-		if confExists {
-			jsonFile, err := os.Open(confFile)
-			if err != nil {
-				return nil, fmt.Errorf("open ovs config file %s error: %v", confFile, err)
-			}
-			defer jsonFile.Close()
-			jsonBytes, err := ioutil.ReadAll(jsonFile)
-			if err != nil {
-				return nil, fmt.Errorf("load ovs config file %s: error: %v", confFile, err)
-			}
-			if err := json.Unmarshal(jsonBytes, flatNetConf); err != nil {
-				return nil, fmt.Errorf("parse ovs config file %s: error: %v", confFile, err)
-			}
-			break
-		}
-	}
-
-	return flatNetConf, nil
-}
-
-func mergeConf(netconf, flatNetConf *netConf) (*netConf, error) {
-	if err := mergo.Merge(netconf, flatNetConf); err != nil {
-		return nil, fmt.Errorf("merge with ovs config file: error: %v", err)
-	}
-	return netconf, nil
-}
-
-func pathExists(path string) (bool, error) {
-	_, err := os.Stat(path)
-	if err == nil {
-		return true, nil
-	}
-	if os.IsNotExist(err) {
-		return false, nil
-	}
-	return false, err
-}
-
-func getOvsConfFiles() []string {
-	return []string{"/etc/kubernetes/cni/net.d/ovs.d/ovs.conf", "/etc/cni/net.d/ovs.d/ovs.conf"}
 }
 
 func generateRandomMac() net.HardwareAddr {
@@ -313,7 +195,7 @@ func refetchIface(iface *current.Interface) error {
 	return nil
 }
 
-func splitVlanIds(trunks []*trunk) ([]uint, error) {
+func splitVlanIds(trunks []*types.Trunk) ([]uint, error) {
 	vlans := make(map[uint]bool)
 	for _, item := range trunks {
 		var minID uint = 0
@@ -374,7 +256,7 @@ func CmdAdd(args *skel.CmdArgs) error {
 		ovnPort = string(envArgs.OvnPort)
 	}
 
-	netconf, err := loadAllNetConf(args.StdinData)
+	netconf, err := config.LoadConf(args.StdinData)
 	if err != nil {
 		return err
 	}
@@ -412,7 +294,7 @@ func CmdAdd(args *skel.CmdArgs) error {
 	defer contNetns.Close()
 
 	var origIfName string
-	if netconf.DeviceID != "" {
+	if sriov.IsOvsHardwareOffloadEnabled(netconf.DeviceID) {
 		origIfName, err = sriov.GetVFLinkName(netconf.DeviceID)
 		if err != nil {
 			return err
@@ -420,12 +302,13 @@ func CmdAdd(args *skel.CmdArgs) error {
 	}
 
 	// Cache NetConf for CmdDel
-	if err = utils.SaveCache(getCRef(args.ContainerID, args.IfName), &cachedNetConf{netconf, origIfName}); err != nil {
+	if err = utils.SaveCache(config.GetCRef(args.ContainerID, args.IfName),
+		&types.CachedNetConf{Netconf: netconf, OrigIfName: origIfName}); err != nil {
 		return fmt.Errorf("error saving NetConf %q", err)
 	}
 
 	var hostIface, contIface *current.Interface
-	if netconf.DeviceID != "" {
+	if sriov.IsOvsHardwareOffloadEnabled(netconf.DeviceID) {
 		// SR-IOV Case
 		hostIface, contIface, err = sriov.SetupSriovInterface(contNetns, args.ContainerID, args.IfName, netconf.MTU, netconf.DeviceID)
 		if err != nil {
@@ -521,7 +404,7 @@ func CmdAdd(args *skel.CmdArgs) error {
 		}
 	}
 
-	return types.PrintResult(result, netconf.CNIVersion)
+	return cnitypes.PrintResult(result, netconf.CNIVersion)
 }
 
 func waitLinkUp(ovsDriver *ovsdb.OvsBridgeDriver, ofPortName string, retryCount, interval int) error {
@@ -574,8 +457,8 @@ func removeOvsPort(ovsDriver *ovsdb.OvsBridgeDriver, portName string) error {
 func CmdDel(args *skel.CmdArgs) error {
 	logCall("DEL", args)
 
-	cRef := getCRef(args.ContainerID, args.IfName)
-	cache, err := loadConfFromCache(cRef)
+	cRef := config.GetCRef(args.ContainerID, args.IfName)
+	cache, err := config.LoadConfFromCache(cRef)
 	if err != nil {
 		// If cmdDel() fails, cached netconf is cleaned up by
 		// the followed defer call. However, subsequence calls
@@ -616,7 +499,7 @@ func CmdDel(args *skel.CmdArgs) error {
 	if args.Netns == "" {
 		// The CNI_NETNS parameter may be empty according to version 0.4.0
 		// of the CNI spec (https://github.com/containernetworking/cni/blob/spec-v0.4.0/SPEC.md).
-		if cache.Netconf.DeviceID != "" {
+		if sriov.IsOvsHardwareOffloadEnabled(cache.Netconf.DeviceID) {
 			// SR-IOV Case - The sriov device is moved into host network namespace when args.Netns is empty.
 			// This happens container is killed due to an error (example: CrashLoopBackOff, OOMKilled)
 			var rep string
@@ -658,7 +541,7 @@ func CmdDel(args *skel.CmdArgs) error {
 
 	// Delete can be called multiple times, so don't return an error if the
 	// device is already removed.
-	if cache.Netconf.DeviceID != "" {
+	if sriov.IsOvsHardwareOffloadEnabled(cache.Netconf.DeviceID) {
 		//  SR-IOV Case
 		err = sriov.ReleaseVF(args, cache.OrigIfName)
 		if err != nil && err == ip.ErrLinkNotFound {
@@ -690,23 +573,4 @@ func CmdCheck(args *skel.CmdArgs) error {
 	logCall("CHECK", args)
 	log.Print("CHECK is not yet implemented, pretending everything is fine")
 	return nil
-}
-
-func loadConfFromCache(cRef string) (*cachedNetConf, error) {
-	netCache := &cachedNetConf{}
-	netConfBytes, err := utils.ReadCache(cRef)
-	if err != nil {
-		return nil, fmt.Errorf("error reading cached NetConf with name %s: %v", cRef, err)
-	}
-
-	if err = json.Unmarshal(netConfBytes, netCache); err != nil {
-		return nil, fmt.Errorf("failed to parse NetConf: %v", err)
-	}
-
-	return netCache, nil
-}
-
-// getCRef unique identifier for a container interface
-func getCRef(cid, podIfName string) string {
-	return strings.Join([]string{cid, podIfName}, "-")
 }
